Move implicit vs. autodiff tests
